feat: Split safe-socket function into its own file
