+++ src/index.js
@@ -5,6 +5,7 @@ const mafmt = require('mafmt')
 const withIs = require('class-is')
 const includes = require('lodash.includes')
 const Connection = require('interface-connection').Connection
+const safe = require('./safe-socket')
 
 const maToUrl = require('./ma-to-url')
 const debug = require('debug')
@@ -30,7 +31,7 @@ class WebSockets {
       }
     })
 
-    const conn = new Connection(socket)
+    const conn = new Connection(safe(socket))
     conn.getObservedAddrs = (cb) => cb(null, [ma])
     conn.close = (cb) => socket.close(cb)
